arrow length fixed to float

+++ VinFletcherArrow/Player.cs
@@ -13,7 +13,7 @@
             countOfArrows = 0;
         }
 
-        public void BuyArrow(Vendor vendor, ArrowheadType arrowheadType, Fletching fletching, int length)
+        public void BuyArrow(Vendor vendor, ArrowheadType arrowheadType, Fletching fletching, float length)
         {
             Arrow? desiredArrow = vendor.GetArrow(arrowheadType, fletching, length);
             float arrowCost = vendor.GetCost(desiredArrow);
